More code refactoring in tests
