+++ tests/chains/documents/test_describe_document_set_chain.py
@@ -12,7 +12,7 @@ from docugami_langchain.chains.documents import (
     SummarizeDocumentChain,
 )
 from docugami_langchain.config import DEFAULT_EXAMPLES_PER_PROMPT
-from tests.conftest import TEST_DATA_DIR, verify_chain_response
+from tests.common import TEST_DATA_DIR, verify_chain_response
 from tests.testdata.dgml_samples.dgml_samples_test_data import (
     DG_SAMPLE_TEST_DATA,
     DGSamplesTestData,
