+++ tests/chains/querying/test_sql_fixup_chain.py
@@ -7,7 +7,7 @@ from langchain_core.embeddings import Embeddings
 from langchain_core.language_models import BaseLanguageModel
 
 from docugami_langchain.chains.querying import SQLFixupChain
-from tests.conftest import TEST_DATA_DIR, is_core_tests_only_mode, verify_chain_response
+from tests.common import TEST_DATA_DIR, is_core_tests_only_mode, verify_chain_response
 
 
 @dataclass
