+++ tests/chains/querying/test_suggested_questions_chain_llm.py
@@ -7,7 +7,7 @@ from langchain_core.language_models import BaseLanguageModel
 
 from docugami_langchain.chains import SuggestedQuestionsChain
 from docugami_langchain.tools.reports import connect_to_excel
-from tests.conftest import TEST_DATA_DIR, verify_chain_response
+from tests.common import TEST_DATA_DIR, verify_chain_response
 from tests.testdata.xlsx.sql_test_data import SQL_TEST_DATA, SQLTestData
 
 SQL_SUGGESTIONS_EXAMPLES_FILE = (
