+++ tests/chains/rag/test_simple_rag_chain.py
@@ -19,14 +19,17 @@ from docugami_langchain.retrievers.mappings import (
     build_doc_maps_from_chunks,
     build_full_doc_summary_mappings,
 )
-from tests.conftest import TEST_DATA_DIR, is_core_tests_only_mode, verify_chain_response
+from tests.common import (
+    EXAMPLES_PATH,
+    RAG_TEST_DGML_DATA_DIR,
+    is_core_tests_only_mode,
+    verify_chain_response,
+)
 
 TEST_QUESTION = "What is the accident number for the incident in madill, oklahoma?"
 TEST_ANSWER_OPTIONS = ["DFW08CA044"]
 
 
-RAG_TEST_DGML_DATA_DIR = TEST_DATA_DIR / "dgml_samples"
-EXAMPLES_PATH = TEST_DATA_DIR / "examples"
 if is_core_tests_only_mode():
     # RAG over fewer files when in core tests mode (to speed things up)
     RAG_TEST_DGML_DATA_DIR = RAG_TEST_DGML_DATA_DIR / "NTSB Aviation Incident Reports"
